Async executor moved to config class
ExecutorConfig class created, optimized imports
