+++ src/main/java/com/example/internshipapp/InternshipAppApplication.java
@@ -3,30 +3,15 @@ package com.example.internshipapp;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.scheduling.annotation.AsyncConfigurerSupport;
 import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
 
 @EnableCaching
 @EnableAsync
 @SpringBootApplication
-public class InternshipAppApplication extends AsyncConfigurerSupport {
+public class InternshipAppApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(InternshipAppApplication.class, args);
-    }
-
-    @Override
-    public Executor getAsyncExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(2);
-        executor.setMaxPoolSize(2);
-        executor.setQueueCapacity(500);
-        executor.setThreadNamePrefix("send-mail");
-        executor.initialize();
-        return executor;
     }
 }
